Ensure "%.ms" and "%p" work for const char*
"%.ms" where m is an integer should read no more than m characters if
the argument type is const char*; the string is not required to be
null terminated in this case so we must be careful not to read off the
end.  The solution is an overloaded function formatCStringTruncate to
handle the special case in a typesafe way.

"%p" should print a pointer, but if passed a const char* the pointer
was derferenced and printed as a string.  To avoid this, we define
special forms of formatValue() for all character string types which
carefully check for the %p conversion.