Simplify implementation of "%p" conversion
Instead of creating multiple specializations of formatValue() for the
special case of character pointers, we now just try to convert all types
to (const void*) when the "%p" conversion is active, and print that.  If
a type can't be so converted, it's printed with operator<< as usual.

Having less specializations of formatValue() makes the code simpler and
shortens the compile time.