Implement variable width and precision arguments
Format specs starting with "%*.*" indicate a variable field width and
precision which come in order in the format string with the actual value
to be formatted coming after.

This change implements support for variable width & precision to enhance
printf() compatibility.  Various internals have been changed, most
notably introducing a FormatIterator class to keep track of the state of
the format string and any variable width/precision arguments inserted.

The interface remains mostly unchanged, though the function
formatValueBasic - which could theoretically be overridden by users -
has been removed.  In hindsight this was just rather complex, and it's
not clear it was necessary since users can override formatValue with
roughly the same results.