Error checking for wchar_t printf() incompatibility
Attempting to format wchar_t* with "%ls" would result in conversion to
void*, and subsequent printing as a pointer - this could result in
unexpected results when converting from a previous usage of traditional
printf().  Changes in this patch make sure that such usage results in a
compile time error to minimize nasty surprises!

Also make the format string parser slightly more pedantic about which
format strings it will accept, by rejecting C99 hexadecimal %a and %A
with TINYFORMAT_ERROR.