Fix clang -Wuninitialized warning
Initialization of FormatList::m_formatters from m_formatterStore isn't
legal in the constructor if we use a std::array, since we can't get a
pointer to the first element of the array without calling a method.

The error is harmless because no member data is accessed, but causes
uninitialized warnings with clang-3.7.  Using a plain old C array avoids
the warning, allows a MSVC workaround to be removed, and avoids an extra
include.