Add format_c, printf_c locale-independent formatting.
The default functions all are dependent on either the global native
locale, or the locale associated with the stream they are outputting
to. This can lead to situations where format("%g",123.45) might return
"123,45" if it happens to be running on a machine with locale "fr_FR",
for example. If that is written to an output file that is later read on
a computer using a locale with '.' as the decimal mark, it could
mis-parse as "123.0", leading to hilarious and tragic results.

This patch adds format_c, printf_c, printfln_c, which force the classic
"C" locale ('.' decimals, among other things). This incurs a performance
penalty -- in particular, when being used on I/O streams it will force a
flush when the locale for the stream is saved and restored -- but it is
the safe thing to do for persistent, portable output.